refactor `distances.py` into `distances` module. 1 file per type!
